add parsing and merging utilities
