--- onepage/cli/__init__.py
@@ -1,5 +1,0 @@
-"""Command-line interface for onepage."""
-
-from .main import main
-
-__all__ = ["main"]
