--- onepage/processing/__init__.py
@@ -1,1 +1,0 @@
-"""Content processing and alignment modules."""
