--- onepage/renderers/__init__.py
@@ -1,1 +1,0 @@
-"""Output renderers for different formats."""
