+++ onepage/translate.py
@@ -3,6 +3,8 @@
 import requests
 from typing import List, Dict, Optional, Tuple
 import time
+import re
+import wikitextparser as wtp
 
 try:
     from langdetect import detect
@@ -12,7 +14,7 @@ except ImportError:
     def detect(text: str) -> str:
         return "unknown"
 
-from ..core.models import Claim
+from .models import Claim
 
 
 class TranslationService:
@@ -205,5 +207,5 @@ class TextCleaner:
         
         # Normalize URLs
         ref_text = re.sub(r'https?://(www\.)?', 'https://', ref_text)
-        
+
         return ref_text
