+++ tests/test_api.py
@@ -3,8 +3,7 @@
 import pytest
 from unittest.mock import Mock, patch
 
-from onepage.api.wikidata import WikidataClient
-from onepage.api.wikipedia import WikipediaClient
+from onepage.api import WikidataClient, WikipediaClient
 
 
 class TestWikidataClient:
