+++ tests/test_models.py
@@ -3,9 +3,14 @@
 import json
 import pytest
 
-from onepage.core.models import (
-    Entity, Claim, Fact, Reference, Provenance, 
-    Section, IntermediateRepresentation
+from onepage.models import (
+    Entity,
+    Claim,
+    Fact,
+    Reference,
+    Provenance,
+    Section,
+    IntermediateRepresentation,
 )
